fixing controller and umlview files to not use extra files
